requirements.txt
openai
requests
openai
requests
python-dotenv
pydub
pyaudio
SpeechRecognition
pygame